--- spoti-friends/src/RealmDatabase.swift
@@ -1,12 +1,0 @@
-import Foundation
-import RealmSwift
-
-class RealmDatabase {
-//    func getRealm() -> Realm {
-//        var realm: Realm;
-//        DispatchQueue.main.async {
-//            realm = try! Realm();
-//        }
-//        return realm;
-//    }
-}
