+++ spoti-friends/src/RootView.swift
@@ -6,6 +6,7 @@ struct RootView: View {
     @EnvironmentObject private var userViewModel: UserViewModel
     
     var body: some View {
+        // Navigate to the appropriate view depending on the user's authorization status
         NavigationStack {
             switch userViewModel.user.authorizationStatus {
             case .unauthenticated:
@@ -22,6 +23,6 @@ struct RootView: View {
 
 struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
-        RootView()
+        RootView().environmentObject(UserViewModel())
     }
 }
