+++ spoti-friends/src/common/utils/Utils.swift
